% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_based.R
\name{conr_rules}
\alias{conr_rules}
\title{Wrapper for tidymodels interface for ConR thresholds.}
\usage{
conr_rules(formula, data)
}
\arguments{
\item{formula}{An object of class \link[stats:formula]{stats::formula} specifying the target column
in relation to the EOO, AOO, and locations column, e.g. y ~ eoo + aoo + locs.}

\item{data}{A dataframe with the target, EOO, AOO, and number of locations.}
}
\description{
Wrapper for tidymodels interface for ConR thresholds.
}
