% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_based.R
\name{conr_thresholds}
\alias{conr_thresholds}
\title{Classify species as threatened using \href{https://github.com/gdauby/conr}{ConR} rules.}
\usage{
conr_thresholds(eoo, aoo, locations)
}
\arguments{
\item{eoo}{The EOO in km^2.}

\item{aoo}{The AOO in km^2.}

\item{locations}{The number of threat-defined locations estimated by ConR.}
}
\value{
classification of whether a species is threatened or not.
}
\description{
\link[ConR:IUCN.eval]{ConR::IUCN.eval} approximates a full classification based on criterion
B, using the EOO, AOO, and an estimate of the number of locations
for a species. This function implements that classification, then
makes the prediction coarser to "threatened" or "not threatened".
}
