% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_based.R
\name{eoo_threshold}
\alias{eoo_threshold}
\title{Classify species as threatened or not based on the IUCN threshold
on EOO (criterion B1).}
\usage{
eoo_threshold(eoo)
}
\arguments{
\item{eoo}{The EOO in km}
}
\value{
classification of whether a species is threatened or not.
}
\description{
Classify species as threatened or not based on the IUCN threshold
on EOO (criterion B1).
}
