% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn.R
\name{iucnn}
\alias{iucnn}
\title{Neural Network classifier to automate occurrence-based conservation assessments}
\usage{
iucnn(
  mode = "classification",
  engine = "keras",
  layers = NULL,
  dropout = NULL,
  epochs = NULL
)
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode.
Only "classification" is allowed.}

\item{engine}{A single character string specifying the engine to use.}

\item{layers}{A string specification of the hidden units in each layer, e.g. "40_20"
for a two-layer network with a 40-unit layer then a 20-unit layer.}

\item{dropout}{A number between 0 (inclusive) and 1 denoting the proportion
of model parameters randomly set to zero during model training.}

\item{epochs}{An integer for the number of training iterations.}
}
\description{
\code{iucnn()} defines a neural network for predicting the conservation status of
species given species-level predictors calculated from occurrence records.
This is an implementation of the \href{https://doi.org/10.1111/ddi.13450}{IUCNN model}
so it works in the tidymodels framework.
}
\details{
Currently only the binary threatened/not threatened classification is implemented.
}
\examples{
parsnip::show_engines("iucnn")

iucnn(layers="40_20", dropout=0.3, epochs=10)

}
