% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnn.R
\name{iucnn_model}
\alias{iucnn_model}
\title{Interface to a neural network model for automated species conservation assessments via keras}
\usage{
iucnn_model(
  x,
  y,
  layers = "30",
  dropout = 0,
  epochs = 30,
  validation_data = NULL,
  save_history = FALSE,
  ...
)
}
\arguments{
\item{x}{A dataframe of matrix of predictors}

\item{y}{A vector (factor or character) of outcome data.}

\item{layers}{A string specification of the number of hidden units in each layer, e.g. "40_20"}

\item{dropout}{The proportion of parameters to set to zero.}

\item{epochs}{An integer for the number of passes through the data.}

\item{validation_data}{A dataframe of data for validation with columns for
predictiors in \code{x} and the outcome \code{y}, or a list of two matrices for the predictors and outcome.}

\item{save_history}{Logical, whether to save the loss values for each epoch during training.}

\item{...}{compilation and fitting arguments to pass to the \code{keras} model.}
}
\value{
A \code{keras} model object.
}
\description{
\code{iucnn_model()} builds a sequential \code{keras} model from a string specification of the units
in each layer of the network. Regularisation is by dropout.
}
