% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{layers}
\alias{layers}
\alias{values_layers}
\title{Parameter objects for occurrence-based automated conservation assessment models.}
\format{
An object of class \code{character} of length 8.
}
\usage{
layers(values = values_layers)

values_layers
}
\arguments{
\item{values}{The possible layer specifications to use.}
}
\value{
A function of class 'qual_param' and 'param'.
}
\description{
\code{layers()} describes a specification of the layers in neural network as
a string with the number of hidden units in each layer separated by an underscore,
e.g "40_20" specifies a two-layer network with 40 units in the first layer and 20
in the second.
}
\examples{
layers(c("10", "40_20", "50_30_10"))

}
\keyword{datasets}
