% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myrcia_predictors}
\alias{myrcia_predictors}
\title{Extinction risk predictors for Myrcia species.}
\format{
A data frame with 666 rows and 18 variables:
\describe{
\item{wcvp_id}{a unique identifier for the species, from \href{https://wcvp.science.kew.org/}{WCVP}},
\item{eoo}{Extent of occurrence (EOO), in km^2},
\item{human_footprint}{a measure of the impact of human activities, in 2009},
\item{human_population}{human population density in 2020, in persons/km^2},
\item{forest_loss}{proportion of a species observed localities that underwent forest loss between 2000-2019},
\item{annual_temperature}{mean average temperature, in degrees-C},
\item{precipitation_driest}{precipitation in the driest quarter, in mm},
\item{elevation}{elevation, in m},
\item{conr_eoo}{EOO, as calculated by \href{https://gdauby.github.io/ConR/}{ConR}, in km^2},
\item{conr_aoo}{Area of occupancy (AOO), as calculated by \href{https://gdauby.github.io/ConR/}{ConR}, in km^2},
\item{conr_locs}{Number of locations estimated by \href{https://gdauby.github.io/ConR/}{ConR}},
\item{conr_obs}{Number of unique observations calculated by \href{https://gdauby.github.io/ConR/}{ConR}},
\item{n_specimens}{Total number of observations},
\item{centroid_latitude}{Latitude of range centroid, in decimal-degrees},
\item{name}{species name, as listed in \href{https://wcvp.science.kew.org/}{WCVP}},
\item{category}{IUCN Red List category, if assessed},
\item{genus}{genus of species, as listed in \href{https://wcvp.science.kew.org/}{WCVP}},
\item{section}{section of species, as detailed in \href{https://doi.org/10.1007/s12225-017-9730-5}{Lucas et al. (2019)}}
}
}
\source{
\url{https://doi.org/10.32942/osf.io/zxq6s}
}
\usage{
myrcia_predictors
}
\description{
A dataset containing predictors to use in automated conservation
methods for species of the genus \emph{Myrcia}. These predictors were prepared for
the paper \href{https://doi.org/10.32942/osf.io/zxq6s}{Evidence-based guidelines for automated assessments of plant species},
and described in detail there. The dataset contains a mixture of assessed and unassessed species.
Assessments were compiled from the \href{https://www.iucnredlist.org/}{IUCN Red List v.2021-3} and predictors were calculated across
observed localities for each species as downloaded from \href{https://www.gbif.org/occurrence/download/0132231-210914110416597}{GBIF}.
}
\keyword{datasets}
