% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_based.R
\name{rule_based}
\alias{rule_based}
\title{Rule based classification}
\usage{
rule_based(mode = "classification", engine = "eoo")
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode.
Only "classification is allowed.}

\item{engine}{A single character string specifying the rule-set engine to use.}
}
\description{
\code{rule_based()} defines a model that uses fixed rules to classify an input.
}
\details{
As the model uses fixed rules, fitting it just creates a model object.
Different sets of rules are implemented as model \emph{engines}.

Information about the class of model fit can be found in \link{rule_based_assessment}.
The engine-specific pages for the different rule sets can be found in:
\itemize{
\item \link{eoo_threshold} (default)
\item \link{conr_thresholds}
}

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.
}
\examples{
parsnip::show_engines("rule_based")

rule_based(mode="classification", engine="conr")

}
