% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_based_assessment.R
\name{rule_based_assessment}
\alias{rule_based_assessment}
\title{A rule-based model.}
\usage{
rule_based_assessment(formula, data, rules)
}
\arguments{
\item{formula}{an object of class \link[stats:formula]{stats::formula}, should specify the target
and all predictors used in the rule function as a linear combination,
e.g. target ~ rule1 + rule2.}

\item{data}{a dataframe containing the variables and target of the rule function.}

\item{rules}{a function implementing the rules-based classification.}
}
\value{
an object class similar to "glm" that can be used with a 'predict' method.
}
\description{
Provides a \link[stats:glm]{stats::glm} type wrapper around user-defined function specifying
rules for a classification function.
}
\examples{

d <- data.frame(
  y=c("apple", "apple", "not apple", "not apple"),
  shape=c("round", "round", "square", "round"),
  colour=c("green", "red", "green", "purple")
)

f <- function(shape, colour) {
  ifelse(shape == "round" & colour \%in\% c("green", "red"), "apple", "not apple")
}

apple_classifier <- rule_based_assessment(y ~ shape + colour, d, f)

predict(apple_classifier, newdata=data.frame(shape="round", colour="indigo"))

}
